Second task is done: project is into modules, webpack is set up
